requirements.txt
django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
dj-database-url==2.1.0
factory-boy==3.3.0
gunicorn==21.2.0
psycopg2-binary==2.9.8